urunle ilgili policy oluşturuldu

+++ app/Http/Controllers/IndexController.php
@@ -6,9 +6,11 @@ use App\Http\Controllers\Controller;
 use App\Models\Urun;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class IndexController extends Controller
 {
+   use AuthorizesRequests;
     public function anasayfaGoster(){
       if(Auth::check()){
         $kullanici=Auth::user();
@@ -22,11 +24,13 @@ $urunler = Urun::where('user_id', $kullaniciId)->with('kategori')->get();
     
     public function sil($id){
    $urun=Urun::findOrFail($id);//ürün var mı kontrol et
+    $this->authorize('delete', $urun);
     $urun->delete();  
     return redirect()->back()->with("basari","ürün basariyla silindi"); 
 }
     public function duzenle($id){
    $urun=Urun::findOrFail($id);
+   $this->authorize("update",$urun);
       return view("duzenle",compact("urun"));
     }
     public function guncelle(Request $request,$id){
